# remote write: reduce remote write egress bytes with new proto format