--- a/prompb/remote.proto
+++ b/prompb/remote.proto
@@ -27,6 +27,17 @@ message WriteRequest {
   repeated prometheus.MetricMetadata metadata = 3 [(gogoproto.nullable) = false];
 }
 
+
+message WriteRequestWithRefs {
+  repeated prometheus.ReducedTimeSeries timeseries = 1 [(gogoproto.nullable) = false];
+  // Cortex uses this field to determine the source of the write request.
+  // We reserve it to avoid any compatibility issues.
+  reserved  2;
+  repeated prometheus.MetricMetadata metadata = 3 [(gogoproto.nullable) = false];
+  // Symbol table for label names/values.
+  map<uint64, string> string_symbol_table = 4 [(gogoproto.nullable) = false];
+}
+
 // ReadRequest represents a remote read request.
 message ReadRequest {
   repeated Query queries = 1;