--- a/prompb/types.proto
+++ b/prompb/types.proto
@@ -54,6 +54,15 @@ message Exemplar {
   int64 timestamp = 3;
 }
 
+message ExemplarRef {
+    // Optional, can be empty.
+    repeated LabelRef labels = 1 [(gogoproto.nullable) = false];
+    double value = 2;
+    // timestamp is in ms format, see model/timestamp/timestamp.go for
+    // conversion from time.Time to Prometheus timestamp.
+    int64 timestamp = 3;
+}
+
 // A native histogram, also known as a sparse histogram.
 // Original design doc:
 // https://docs.google.com/document/d/1cLNv3aufPZb3fNfaJgdaRBZsInZKKIHo9E6HinJVbpM/edit
@@ -127,6 +136,21 @@ message TimeSeries {
   repeated Sample samples       = 2 [(gogoproto.nullable) = false];
   repeated Exemplar exemplars   = 3 [(gogoproto.nullable) = false];
   repeated Histogram histograms = 4 [(gogoproto.nullable) = false];
+}
+
+// TimeSeries represents samples and labels for a single time series.
+message ReducedTimeSeries {
+  // For a timeseries to be valid, and for the samples and exemplars
+  // to be ingested by the remote system properly, the labels field is required.
+  repeated LabelRef labels         = 1 [(gogoproto.nullable) = false];
+  repeated Sample samples       = 2 [(gogoproto.nullable) = false];
+  repeated ExemplarRef exemplars   = 3 [(gogoproto.nullable) = false];
+  repeated Histogram histograms = 4 [(gogoproto.nullable) = false];
+}
+
+message LabelRef {
+  uint64 name_ref = 1;
+  uint64 value_ref = 2;
 }
 
 message Label {