--- a/scripts/genproto.sh
+++ b/scripts/genproto.sh
@@ -10,8 +10,8 @@ if ! [[ "$0" =~ "scripts/genproto.sh" ]]; then
 	exit 255
 fi
 
-if ! [[ $(protoc --version) =~ "3.15.8" ]]; then
-	echo "could not find protoc 3.15.8, is it installed + in PATH?"
+if ! [[ $(protoc --version) =~ "3.21.12" ]]; then
+	echo "could not find protoc 3.21.12, is it installed + in PATH?"
 	exit 255
 fi