--- a/storage/remote/write.go
+++ b/storage/remote/write.go
@@ -17,6 +17,8 @@ import (
 	"context"
 	"fmt"
 	"math"
+	"math/rand"
+	"strconv"
 	"sync"
 	"time"
 
@@ -65,6 +67,7 @@ type WriteStorage struct {
 	externalLabels    labels.Labels
 	dir               string
 	queues            map[string]*QueueManager
+	writeReducedProto bool
 	samplesIn         *ewmaRate
 	flushDeadline     time.Duration
 	interner          *pool
@@ -76,12 +79,13 @@ type WriteStorage struct {
 }
 
 // NewWriteStorage creates and runs a WriteStorage.
-func NewWriteStorage(logger log.Logger, reg prometheus.Registerer, dir string, flushDeadline time.Duration, sm ReadyScrapeManager) *WriteStorage {
+func NewWriteStorage(logger log.Logger, reg prometheus.Registerer, dir string, flushDeadline time.Duration, sm ReadyScrapeManager, writeReducedProto bool) *WriteStorage {
 	if logger == nil {
 		logger = log.NewNopLogger()
 	}
 	rws := &WriteStorage{
 		queues:            make(map[string]*QueueManager),
+		writeReducedProto: writeReducedProto,
 		watcherMetrics:    wlog.NewWatcherMetrics(reg),
 		liveReaderMetrics: wlog.NewLiveReaderMetrics(reg),
 		logger:            logger,
@@ -152,17 +156,22 @@ func (rws *WriteStorage) ApplyConfig(conf *config.Config) error {
 		if rwConf.Name != "" {
 			name = rwConf.Name
 		}
-
-		c, err := NewWriteClient(name, &ClientConfig{
-			URL:              rwConf.URL,
-			Timeout:          rwConf.RemoteTimeout,
-			HTTPClientConfig: rwConf.HTTPClientConfig,
-			SigV4Config:      rwConf.SigV4Config,
-			Headers:          rwConf.Headers,
-			RetryOnRateLimit: rwConf.QueueConfig.RetryOnRateLimit,
-		})
-		if err != nil {
-			return err
+		var c WriteClient
+		if rwConf.URL.String() == "fake" {
+			f := "fake" + strconv.Itoa(rand.Intn(100))
+			c = NewTestClient(f, f)
+		} else {
+			c, err = NewWriteClient(name, &ClientConfig{
+				URL:              rwConf.URL,
+				Timeout:          rwConf.RemoteTimeout,
+				HTTPClientConfig: rwConf.HTTPClientConfig,
+				SigV4Config:      rwConf.SigV4Config,
+				Headers:          rwConf.Headers,
+				RetryOnRateLimit: rwConf.QueueConfig.RetryOnRateLimit,
+			})
+			if err != nil {
+				return err
+			}
 		}
 
 		queue, ok := rws.queues[hash]
@@ -196,6 +205,7 @@ func (rws *WriteStorage) ApplyConfig(conf *config.Config) error {
 			rws.scraper,
 			rwConf.SendExemplars,
 			rwConf.SendNativeHistograms,
+			rws.writeReducedProto,
 		)
 		// Keep track of which queues are new so we know which to start.
 		newHashes = append(newHashes, hash)