@cstyan 2023-02-21
To extend this, we could allow the option of using alternative encoding/compression types. For example, with zstd we can achieve another 25% reduction in the end compressed bytes size but require double the amount of time to produce those compressed bytes:
```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/storage/remote
cpu: AMD Ryzen 9 5950X 16-Core Processor
                            │ snappy.txt  │               zstd.txt               │
                            │   sec/op    │   sec/op     vs base                 │
BuildReducedWriteRequest-32   2.459m ± 1%   5.098m ± 2%  +107.29% (p=0.000 n=10)

                            │    snappy.txt     │                 zstd.txt                  │
                            │ compressedSize/op │ compressedSize/op  vs base                │
BuildReducedWriteRequest-32         238.2k ± 0%         174.1k ± 0%  -26.94% (p=0.000 n=10)

                            │  snappy.txt  │              zstd.txt               │
                            │     B/op     │     B/op      vs base               │
BuildReducedWriteRequest-32   1.188Mi ± 0%   1.188Mi ± 0%  -0.00% (p=0.000 n=10)

                            │ snappy.txt │            zstd.txt            │
                            │ allocs/op  │ allocs/op   vs base            │
BuildReducedWriteRequest-32   3.000 ± 0%   3.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```