@gouthamve 2023-07-26
OTLP has done some experiments in this space and they saw a lot of benefit in compression and speed by trying an arrow based implementation. 

See https://github.com/open-telemetry/oteps/blob/main/text/0156-columnar-encoding.md and https://github.com/f5/otel-arrow-adapter/blob/main/docs/benchmarks.md

I am not sure we should adopt it, but it might be something to evaluate.

Further, my naive benchmark using avalanche data saw that OTLP (non-arrow) is more efficient on the compression. See: https://github.com/gouthamve/otlp-prw-compare